requirements.txt
opencv-python
mediapipe
numpy
pyvista
pyaml
screeninfo